pyproject.toml
[tool.poetry]
name = "k8s-jobs"
version = "0.1.5"
description = "Async Job Manager + AWS Batch Replacement for K8s"
authors = ["Daniel Tahara <dktahara@gmail.com>"]
repository = "https://github.com/danieltahara/k8s-jobs"
readme = "README.rst"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.6,<3.8"
kubernetes = "^10.0"
jinja2 = "^2.10"
dataclasses = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest-cov = "^2.7"
flask = "^1.1"
gitpython = "^3.0"
pytest-timeout = "^1.3"

[tool.black]
line-length = 88
target-version = ['py36', 'py37']

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
